cling

+++ custom_features/department_specific/business_hs.js
@@ -241,8 +241,8 @@
                     let courseCount = this.courseGrades.length;
                     for (let c = 0; c < courseCount; c++) {
                       let courseData = this.courseGrades[c];
-                      console.log(parseInt(courseData['grade']));
-                      coursePointsTotal += parseInt(courseData['grade']);
+                      console.log(parseFloat(courseData['grade']));
+                      coursePointsTotal += parseFloat(courseData['grade']);
                     }
                     let average = Math.round((coursePointsTotal / courseCount) * 10) * .1
                     console.log(average);
